src/pyproject.toml
[tool.poetry]
name = "ngru.API"
version = "0.1.0"
description = "Simple API for recruitment purposes"
authors = ["Jedrzej Szadejko <jedrzej@szadejko.pl>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
flake8 = "^3.8.4"
isort = "^5.7.0"
black = "^20.8b1"
pytest = "^6.2.2"
psycopg2 = "^2.8.6"
pytest-cov = "^2.11.1"
pytest-django = "^4.1.0"
python-decouple = "^3.4"
requests = "^2.25.1"
gunicorn = "^20.0.4"
uvicorn = {extras = ["standard"], version = "^0.12.0"}

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 95
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''


[tool.isort]
force_single_line = true
import_heading_stdlib="Builtins"
import_heading_firstparty="Local"
import_heading_thirdparty="3rd party"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
